Fix type of argument to copy_remote()
The @uid argument should be uint32_t not unsigned long.

+++ imap/sync_client.c
@@ -804,7 +804,7 @@ done:
     return r;
 }
 
-static int copy_remote(struct mailbox *mailbox, unsigned long uid,
+static int copy_remote(struct mailbox *mailbox, uint32_t uid,
 		       struct dlist *kr)
 {
     struct index_record record;
